# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Personas y Aulas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clases": {
            "get": {
                "tags": [
                    "Clases"
                ],
                "summary": "Listar clases",
                "operationId": "6b51d92341941d3b22829adbd68dfa7b",
                "responses": {
                    "200": {
                        "description": "Lista de clases"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clases"
                ],
                "summary": "Crear clase",
                "operationId": "f55bdd808ee3b056f884b4b179a328f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Clase creada"
                    }
                }
            }
        },
        "/api/clases/{id}": {
            "put": {
                "tags": [
                    "Clases"
                ],
                "summary": "Actualizar clase",
                "operationId": "82a086a91ab3bf2e109e92684fadbb34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la clase",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Clase actualizada"
                    },
                    "404": {
                        "description": "Clase no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clases"
                ],
                "summary": "Eliminar clase",
                "operationId": "bb3f338fc77b4722fb422a2063159a6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la clase",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Clase eliminada"
                    },
                    "404": {
                        "description": "Clase no encontrada"
                    }
                }
            }
        },
        "/api/clases_estudiantes": {
            "get": {
                "tags": [
                    "Clases Estudiantes"
                ],
                "summary": "Listar clases_estudiantes",
                "operationId": "c1be90931049e9080eba53c6a9a2002c",
                "responses": {
                    "200": {
                        "description": "Lista de clases_estudiantes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clases Estudiantes"
                ],
                "summary": "Crear clase_estudiante",
                "operationId": "b7e81ebd3589a81cb079a53b6db84b21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_clase",
                                    "id_estudiante"
                                ],
                                "properties": {
                                    "id_clase": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_estudiante": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Clase_Estudiante creada"
                    }
                }
            }
        },
        "/api/clases_estudiantes/{id}": {
            "put": {
                "tags": [
                    "Clases Estudiantes"
                ],
                "summary": "Actualizar clase_estudiante",
                "operationId": "da5a1c5d50a5e82d56fc5934bb99c74d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la clase_estudiante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_clase",
                                    "id_estudiante"
                                ],
                                "properties": {
                                    "id_clase": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_estudiante": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Clase_Estudiante actualizada"
                    },
                    "404": {
                        "description": "Clase_Estudiante no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clases Estudiantes"
                ],
                "summary": "Eliminar clase_estudiante",
                "operationId": "859e8fb4c7e79fb3f223a0f24fc0b594",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la clase_estudiante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Clase_Estudiante eliminada"
                    },
                    "404": {
                        "description": "Clase_Estudiante no encontrada"
                    }
                }
            }
        },
...
    },
...
}